third_party/blink/web_tests/svg/dynamic-updates/SVGFEMorphologyElement-dom-operator-attr-expected.txt

SVG 1.1 dynamic update tests
Erode radius = 4

Tests dynamic updates of the 'operator' attribute of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS morphologyElement.getAttribute('operator') is "dilate"
PASS morphologyElement.getAttribute('operator') is "erode"
PASS successfullyParsed is true

TEST COMPLETE
